+++ onsiteprint-plugin/assets/js/inc/modal/modal.js
@@ -1,0 +1,24 @@
+/* ------------------------------------------------------------------------
+ #  JS Part Name: Modal
+ *  Functions related to the Modal Window.
+ ?  Updated: 2025-08-06 - 04:13 (Y:m:d - H:i)
+ ?  Info: Added new Script, Modal.
+---------------------------------------------------------------------------
+ #  TABLE OF CONTENTS:
+---------------------------------------------------------------------------
+
+    1.  Change Modal Content
+    
+    2.  Modal Toggle Listener
+
+---------------------------------------------------------------------------
+ #  1. Change Modal Content
+--------------------------------------------------------------------------- */
+import { opChangeModalContent } from './change-modal-content.js'
+export  { opChangeModalContent as changeContent }
+
+/* ------------------------------------------------------------------------
+ #  2. Modal Toggle Listener
+--------------------------------------------------------------------------- */
+import { opModalToggleListener } from './toggle-modal-listener.js'
+export  { opModalToggleListener as toggleListener }